feat: allow to specifcy `since` and select repository or org
To get the changelog for the current repository since a specific date
use the `--since` parameter:
```
gh terasology changelog --since 2021-12-01
```

To select a different single repository use the `--repo` (or `-R`) flag:
```
gh terasology changelog --repo MovingBlocks/TerasologyLauncher
```

To compile a changelog for all repositories of an organization run the
following. This requires an explicit `--since` flag as the "latest
release" cannot be derived in this case.
```
gh terasology changelog --repo Terasology --since 2021-12-01
```